Moved personalized banners to their own components
